requirements.txt
fastapi
uvicorn
python-multipart
pillow
openai>=1.0.0